Move terminal-related APIs from node-core-library to the terminal package.
