+++ build-tests/install-test-workspace/package.json
@@ -10,6 +10,7 @@
   "devDependencies": {
     "@microsoft/rush-lib": "workspace:*",
     "@rushstack/node-core-library": "workspace:*",
-    "@rushstack/rush-sdk": "workspace:*"
+    "@rushstack/rush-sdk": "workspace:*",
+    "@rushstack/terminal": "workspace:*"
   }
 }
