+++ libraries/node-core-library/src/index.ts
@@ -8,7 +8,6 @@
  */
 
 export { AlreadyReportedError } from './AlreadyReportedError';
-export { AnsiEscape, IAnsiEscapeConvertForTestsOptions } from './Terminal/AnsiEscape';
 export { Async, AsyncQueue, IAsyncParallelismOptions, IRunWithRetriesOptions } from './Async';
 export { Brand } from './PrimitiveTypes';
 export { FileConstants, FolderConstants } from './Constants';
@@ -102,21 +101,4 @@ export { FileWriter, IFileWriterFlags } from './FileWriter';
 export { LegacyAdapters, LegacyCallback } from './LegacyAdapters';
 export { StringBuilder, IStringBuilder } from './StringBuilder';
 export { ISubprocessOptions, SubprocessTerminator } from './SubprocessTerminator';
-export { ITerminal } from './Terminal/ITerminal';
-export { Terminal } from './Terminal/Terminal';
-export { Colors, IColorableSequence, ColorValue, TextAttribute } from './Terminal/Colors';
-export { ITerminalProvider, TerminalProviderSeverity } from './Terminal/ITerminalProvider';
-export { ConsoleTerminalProvider, IConsoleTerminalProviderOptions } from './Terminal/ConsoleTerminalProvider';
-export {
-  StringBufferTerminalProvider,
-  IStringBufferOutputOptions
-} from './Terminal/StringBufferTerminalProvider';
-export {
-  PrefixProxyTerminalProvider,
-  IPrefixProxyTerminalProviderOptions,
-  IDynamicPrefixProxyTerminalProviderOptions,
-  IPrefixProxyTerminalProviderOptionsBase,
-  IStaticPrefixProxyTerminalProviderOptions
-} from './Terminal/PrefixProxyTerminalProvider';
-export { TerminalWritable, ITerminalWritableOptions } from './Terminal/TerminalWritable';
 export { TypeUuid } from './TypeUuid';
