+++ libraries/rush-lib/src/logic/setup/KeyboardLoop.ts
@@ -6,7 +6,7 @@ import * as readline from 'readline';
 import * as process from 'process';
 import { AlreadyReportedError, InternalError } from '@rushstack/node-core-library';
 
-// TODO: Integrate these into the AnsiEscape API in @rushstack/node-core-library
+// TODO: Integrate these into the AnsiEscape API in @rushstack/terminal
 // As part of that work we should generalize the "Colors" API to support more general
 // terminal escapes, and simplify the interface for that API.
 const ANSI_ESCAPE_SHOW_CURSOR: string = '\u001B[?25l';
