+++ libraries/terminal/src/ConsoleTerminalProvider.ts
