+++ libraries/terminal/src/ITerminal.ts
