+++ libraries/terminal/src/ITerminalProvider.ts
