+++ libraries/terminal/src/MockWritable.ts
@@ -1,9 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+import { AnsiEscape } from './AnsiEscape';
 import type { ITerminalChunk } from './ITerminalChunk';
 import { TerminalWritable } from './TerminalWritable';
-import { AnsiEscape } from '@rushstack/node-core-library';
 
 /**
  * A {@link TerminalWritable} subclass for use by unit tests.
