+++ libraries/terminal/src/PrefixProxyTerminalProvider.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+import { Text } from '@rushstack/node-core-library';
 import type { ITerminalProvider, TerminalProviderSeverity } from './ITerminalProvider';
-import { Text } from '../Text';
 
 /**
  * @beta
