+++ libraries/terminal/src/PrintUtilities.ts
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import type { ITerminal } from '@rushstack/node-core-library';
+import type { ITerminal } from './ITerminal';
 
 /**
  * A sensible fallback column width for consoles.
