+++ libraries/terminal/src/StringBufferTerminalProvider.ts
@@ -1,9 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
+import { StringBuilder, Text } from '@rushstack/node-core-library';
 import { type ITerminalProvider, TerminalProviderSeverity } from './ITerminalProvider';
-import { StringBuilder } from '../StringBuilder';
-import { Text } from '../Text';
 import { AnsiEscape } from './AnsiEscape';
 
 /**
