+++ libraries/terminal/src/test/Colors.test.ts
