+++ libraries/terminal/src/test/PrefixProxyTerminalProvider.test.ts
