+++ libraries/terminal/src/test/PrintUtilities.test.ts
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { StringBufferTerminalProvider, Terminal } from '@rushstack/node-core-library';
-
+import { StringBufferTerminalProvider } from '../StringBufferTerminalProvider';
+import { Terminal } from '../Terminal';
 import { PrintUtilities } from '../PrintUtilities';
 
 describe(PrintUtilities.name, () => {
