+++ libraries/terminal/src/test/Terminal.test.ts
