+++ libraries/terminal/src/test/createColorGrid.ts
@@ -5,7 +5,7 @@
  * This file is a little program that prints all of the colors to the console
  */
 
-import { Colors, type IColorableSequence } from '../../index';
+import { Colors, type IColorableSequence } from '../index';
 
 export function createColorGrid(
   attributeFunction?: (text: string | IColorableSequence) => IColorableSequence
